backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
azure-core==1.35.0
azure-storage-blob==12.26.0
base58==2.1.1
bitarray==3.7.1
boto3==1.40.21
botocore==1.40.21
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
ckzg==2.1.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
cytoolz==1.0.1
dnspython==1.16.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
hexbytes==1.3.1
idna==3.10
ipfshttpclient==0.7.0
isodate==0.7.2
jmespath==1.0.1
motor==3.7.1
multiaddr==0.0.9
multidict==6.6.4
netaddr==1.3.0
parsimonious==0.10.0
propcache==0.3.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyunormalize==16.0.0
pywin32==311
regex==2025.8.29
requests==2.32.5
rlp==4.1.0
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
toolz==1.0.0
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
varint==1.0.2
web3==7.13.0
websockets==15.0.1
yarl==1.20.1
